+++ DebugEngine/Exec/DebuggerProxy.cpp
@@ -263,9 +263,9 @@ namespace MagoCore
     HRESULT DebuggerProxy::ReadMemory( 
         IProcess* process, 
         Address address,
-        SIZE_T length, 
-        SIZE_T& lengthRead, 
-        SIZE_T& lengthUnreadable, 
+        uint32_t length, 
+        uint32_t& lengthRead, 
+        uint32_t& lengthUnreadable, 
         uint8_t* buffer )
     {
         // call it directly for performance (it gets called a lot for stack walking)
@@ -276,8 +276,8 @@ namespace MagoCore
     HRESULT DebuggerProxy::WriteMemory( 
         IProcess* process, 
         Address address,
-        SIZE_T length, 
-        SIZE_T& lengthWritten, 
+        uint32_t length, 
+        uint32_t& lengthWritten, 
         uint8_t* buffer )
     {
         HRESULT             hr = S_OK;
